fix ssl

+++ upload/2delete.sh
@@ -1,21 +1,30 @@
 #!/bin/bash
-domain=$1
+path=$1
 
-PREFIX=$(echo "${domain}" | sed 's/\.//g')
+
+if [ -z "$1" ]; then
+  echo "Error: masukkan domain"
+  exit 1
+fi
+
+
+PREFIX=$(echo "${path}" | sed 's/\.//g')
 
 docker container stop $(docker container ls -q --filter name=${PREFIX}_*)
 docker container rm $(docker ps -a -q --filter name=${PREFIX}_*)
 docker network rm $(docker network ls -q --filter name=${PREFIX}_*)
 docker volume prune -f
-sudo /usr/sbin/userdel --remove $domain
-sudo rm -rf /var/spool/mail/$domain
+sudo userdel -r $path
+#sudo rm -rf /var/spool/mail/$path
 sudo quotacheck -ugmf /home
 echo "Docker dan user dihapus"
 
+echo "Hapus reverse proxy"
 user="root"
 server="103.102.153.32"
 
-sudo ssh "$user@$server" "rm -f /etc/nginx/conf.d/$domain.conf"
+sudo ssh "$user@$server" "rm /etc/nginx/conf.d/$path.conf"
+sudo ssh "$user@$server" "rm -rf /home/$path/"
 sudo ssh "$user@$server" "systemctl restart nginx"
 echo "reverse proxy dihapus"
 exit 1
